+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # CHANGELOG
 All notable changes to this project will be documented in this file. To know better on how to write and maintain a changelog, refer to [this link](https://keepachangelog.com/en/1.0.0/).
 
+## [1.3.0]
+
+### Added
+
+- New error `ProviderOffline` when a provider is unavailable.
+
+### Changed
+
+- Updated the underlying iOS SDK to version [2.5.0](https://github.com/TrueLayer/TrueLayer-iOS-SDK/releases/tag/2.5.0).
+- Updated the underlying Android SDK to version [2.4.0](https://docs.truelayer.com/docs/android-sdk-release-history).
+
+### Fixed
+
+- Error when running version 1.1.0 ([#35](https://github.com/TrueLayer/truelayer-react-native-sdk/issues/35)).
+
 ## [1.2.0]
 
 ### Added